Put the fitbit device to sleep after finishing the sync
This is a fix up of Ben Smithurst's pull request
(https://github.com/qdot/libfitbit/pull/20) so that it applies to the
current fitbit tree.  His original commit request read:

    From reading the logs on windows it appears that the windows
    client sends a final command at the end to put the tracker to
    sleep for the 15 minute sync interval. This change appears to have
    the desired effect of making libfitbit do the same.

After experimenting with this change, I find it's an important thing
to do for two reasons.

1) It definitely improves the battery lifetime of the fitbit
device, in a very noticeable way.

2) If there is multiple people within range of the base station (say,
if there is someone else also using a fitbit on the walkstation right
next to yours), it's important to put the device to sleep so that the
next time the fitbit client runs, the client can talk to the other
device.  Otherwise, the fitbit device with the lower serial number
appears to always "win" and if you have the newer fitbit, you can ever
get the information sent back to the fitbit servers.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>