Immediately restart looking for a new client after finishing a cycle
Now that we are putting the fitbit device to sleep for 15 minutes, it
is not necessary for the fitbit_client.py script to sleep for 15
minutes before restarting the cycle.

If there are multiple devices in range of the base station, this is
important so that the fitbit client can quickly collect information
from both devices.  It is also useful in the situation where someone
wanders by the base station software while it is running; by not
making the base station software be quiscient for 15 minutes between
polls, it makes it more likely that we will collect information from a
fitbit client which is near us, and send that information back to the
fitbit servers.

(This is especially handy if there are going to be large number of
users at a site all using fitbits as part of a employee fitness
program.)

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>